+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- API to serialize KeyRings to binary data:
+	```go
+	func (keyRing *KeyRing) Serialize() ([]byte, error)
+	```
+- API to parse KeyRings from binary data:
+	```go
+	func NewKeyRingFromBinary(binKeys []byte) (*KeyRing, error)
+	```
+
 ## [3.0.0-alpha.1] 2024-03-20
 ### Added
 - Allow to override algorithm in key generation